pyproject.toml
[project]
name = "automata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flet>=0.24.1",
    "pillow>=10.4.0",
]
